refactor: changing container in listitle
